sync webrtc-src@m94

+++ mediasoup-client/deps/webrtc/src/system_wrappers/include/ntp_time.h
@@ -62,51 +62,59 @@ inline bool operator!=(const NtpTime& n1, const NtpTime& n2) {
   return !(n1 == n2);
 }
 
-// Converts |int64_t| milliseconds to Q32.32-formatted fixed-point seconds.
+// Converts `int64_t` milliseconds to Q32.32-formatted fixed-point seconds.
 // Performs clamping if the result overflows or underflows.
 inline int64_t Int64MsToQ32x32(int64_t milliseconds) {
-  // TODO(bugs.webrtc.org/10893): Change to use |rtc::saturated_cast| once the
+  // TODO(bugs.webrtc.org/10893): Change to use `rtc::saturated_cast` once the
   // bug has been fixed.
   double result =
       std::round(milliseconds * (NtpTime::kFractionsPerSecond / 1000.0));
 
-  if (result <= std::numeric_limits<int64_t>::min()) {
+  // Explicitly cast values to double to avoid implicit conversion warnings
+  // The conversion of the std::numeric_limits<int64_t>::max() triggers
+  // -Wimplicit-int-float-conversion warning in clang 10.0.0 without explicit
+  // cast
+  if (result <= static_cast<double>(std::numeric_limits<int64_t>::min())) {
     return std::numeric_limits<int64_t>::min();
   }
 
-  if (result >= std::numeric_limits<int64_t>::max()) {
+  if (result >= static_cast<double>(std::numeric_limits<int64_t>::max())) {
     return std::numeric_limits<int64_t>::max();
   }
 
   return rtc::dchecked_cast<int64_t>(result);
 }
 
-// Converts |int64_t| milliseconds to UQ32.32-formatted fixed-point seconds.
+// Converts `int64_t` milliseconds to UQ32.32-formatted fixed-point seconds.
 // Performs clamping if the result overflows or underflows.
 inline uint64_t Int64MsToUQ32x32(int64_t milliseconds) {
-  // TODO(bugs.webrtc.org/10893): Change to use |rtc::saturated_cast| once the
+  // TODO(bugs.webrtc.org/10893): Change to use `rtc::saturated_cast` once the
   // bug has been fixed.
   double result =
       std::round(milliseconds * (NtpTime::kFractionsPerSecond / 1000.0));
 
-  if (result <= std::numeric_limits<uint64_t>::min()) {
+  // Explicitly cast values to double to avoid implicit conversion warnings
+  // The conversion of the std::numeric_limits<int64_t>::max() triggers
+  // -Wimplicit-int-float-conversion warning in clang 10.0.0 without explicit
+  // cast
+  if (result <= static_cast<double>(std::numeric_limits<uint64_t>::min())) {
     return std::numeric_limits<uint64_t>::min();
   }
 
-  if (result >= std::numeric_limits<uint64_t>::max()) {
+  if (result >= static_cast<double>(std::numeric_limits<uint64_t>::max())) {
     return std::numeric_limits<uint64_t>::max();
   }
 
   return rtc::dchecked_cast<uint64_t>(result);
 }
 
-// Converts Q32.32-formatted fixed-point seconds to |int64_t| milliseconds.
+// Converts Q32.32-formatted fixed-point seconds to `int64_t` milliseconds.
 inline int64_t Q32x32ToInt64Ms(int64_t q32x32) {
   return rtc::dchecked_cast<int64_t>(
       std::round(q32x32 * (1000.0 / NtpTime::kFractionsPerSecond)));
 }
 
-// Converts UQ32.32-formatted fixed-point seconds to |int64_t| milliseconds.
+// Converts UQ32.32-formatted fixed-point seconds to `int64_t` milliseconds.
 inline int64_t UQ32x32ToInt64Ms(uint64_t q32x32) {
   return rtc::dchecked_cast<int64_t>(
       std::round(q32x32 * (1000.0 / NtpTime::kFractionsPerSecond)));
